+++ stdm/composer/custom_layout_items.py
@@ -10,204 +10,18 @@
 """
 Contains custom layout item types
 """
-from qgis.PyQt.QtCore import (
-    QCoreApplication,
-    QPointF
-)
-from qgis.PyQt.QtGui import QPolygonF
+
 from qgis.core import (
-    QgsLayoutItemRegistry,
-    QgsLayoutItemAbstractMetadata,
-    QgsLayoutItemPolyline,
-    QgsLayoutItemLabel,
-    QgsLayoutItemMap,
-    QgsLayoutItemPicture,
-    QgsLayoutItemAttributeTable,
     QgsApplication,
-    QgsLayoutMultiFrameAbstractMetadata
 )
 
-from stdm.ui.gui_utils import GuiUtils
-
-BASE_ITEM_TYPE = QgsLayoutItemRegistry.PluginItem + 2337
-STDM_LINE_ITEM_TYPE = BASE_ITEM_TYPE + 0
-STDM_DATA_LABEL_ITEM_TYPE = BASE_ITEM_TYPE + 1
-STDM_DATA_TABLE_ITEM_TYPE = BASE_ITEM_TYPE + 2
-STDM_MAP_ITEM_TYPE = BASE_ITEM_TYPE + 3
-STDM_PHOTO_ITEM_TYPE = BASE_ITEM_TYPE + 4
-STDM_CHART_ITEM_TYPE = BASE_ITEM_TYPE + 5
-STDM_QR_ITEM_TYPE = BASE_ITEM_TYPE + 6
-
-
-class StdmLineLayoutItem(QgsLayoutItemPolyline):
-
-    def __init__(self, layout):
-        super().__init__(layout)
-        self._size = None
-        self._point = None
-
-    def type(self):
-        return STDM_LINE_ITEM_TYPE
-
-    def icon(self):
-        return GuiUtils.get_icon('line.png')
-
-    def attemptResize(self, size, includesFrame):
-        if not self.nodes():
-            self._size = size
-            self._set_initial_line()
-        else:
-            super().attemptResize(size, includesFrame)
-
-    def attemptMove(self, point, useReferencePoint, includesFrame, page):
-        if not self.nodes():
-            self._point = point
-            self._set_initial_line()
-        else:
-            super().attemptMove(point, useReferencePoint, includesFrame, page)
-
-    def _set_initial_line(self):
-        if self._point is not None and self._size is not None:
-            self.setNodes(QPolygonF([QPointF(self._point.x(), self._point.y()),
-                                     QPointF(self._point.x() + self._size.width(),
-                                             self._point.y() + self._size.height())]))
-            self._point = None
-            self._size = None
-
-
-class StdmLineLayoutItemMetadata(QgsLayoutItemAbstractMetadata):
-
-    def __init__(self):
-        super().__init__(STDM_LINE_ITEM_TYPE, QCoreApplication.translate('StdmItems', 'STDM Line'))
-
-    def createItem(self, layout):
-        return StdmLineLayoutItem(layout)
-
-
-class StdmDataLabelLayoutItem(QgsLayoutItemLabel):
-
-    def __init__(self, layout):
-        super().__init__(layout)
-
-    def type(self):
-        return STDM_DATA_LABEL_ITEM_TYPE
-
-    def icon(self):
-        return GuiUtils.get_icon('db_field.png')
-
-
-class StdmDataLabelLayoutItemMetadata(QgsLayoutItemAbstractMetadata):
-
-    def __init__(self):
-        super().__init__(STDM_DATA_LABEL_ITEM_TYPE, QCoreApplication.translate('StdmItems', 'STDM Data Label'))
-
-    def createItem(self, layout):
-        return StdmDataLabelLayoutItem(layout)
-
-
-class StdmTableLayoutItem(QgsLayoutItemAttributeTable):
-
-    def __init__(self, layout):
-        super().__init__(layout)
-
-    def type(self):
-        return STDM_DATA_TABLE_ITEM_TYPE
-
-    def icon(self):
-        return GuiUtils.get_icon('composer_table.png')
-
-
-class StdmTableLayoutItemMetadata(QgsLayoutMultiFrameAbstractMetadata):
-
-    def __init__(self):
-        super().__init__(STDM_DATA_TABLE_ITEM_TYPE, QCoreApplication.translate('StdmItems', 'STDM Data Table'))
-
-    def createMultiFrame(self, layout):
-        return StdmTableLayoutItem(layout)
-
-
-class StdmMapLayoutItem(QgsLayoutItemMap):
-
-    def __init__(self, layout):
-        super().__init__(layout)
-
-    def type(self):
-        return STDM_MAP_ITEM_TYPE
-
-    def icon(self):
-        return GuiUtils.get_icon('add_map.png')
-
-
-class StdmMapLayoutItemMetadata(QgsLayoutItemAbstractMetadata):
-
-    def __init__(self):
-        super().__init__(STDM_MAP_ITEM_TYPE, QCoreApplication.translate('StdmItems', 'STDM Map'))
-
-    def createItem(self, layout):
-        return StdmMapLayoutItem(layout)
-
-
-class StdmPhotoLayoutItem(QgsLayoutItemPicture):
-
-    def __init__(self, layout):
-        super().__init__(layout)
-
-    def type(self):
-        return STDM_PHOTO_ITEM_TYPE
-
-    def icon(self):
-        return GuiUtils.get_icon('photo_24.png')
-
-
-class StdmPhotoLayoutItemMetadata(QgsLayoutItemAbstractMetadata):
-
-    def __init__(self):
-        super().__init__(STDM_PHOTO_ITEM_TYPE, QCoreApplication.translate('StdmItems', 'STDM Photo'))
-
-    def createItem(self, layout):
-        return StdmPhotoLayoutItem(layout)
-
-
-class StdmChartLayoutItem(QgsLayoutItemPicture):
-
-    def __init__(self, layout):
-        super().__init__(layout)
-
-    def type(self):
-        return STDM_CHART_ITEM_TYPE
-
-    def icon(self):
-        return GuiUtils.get_icon('chart.png')
-
-
-class StdmChartLayoutItemMetadata(QgsLayoutItemAbstractMetadata):
-
-    def __init__(self):
-        super().__init__(STDM_CHART_ITEM_TYPE, QCoreApplication.translate('StdmItems', 'STDM Chart'))
-
-    def createItem(self, layout):
-        return StdmChartLayoutItem(layout)
-
-
-class StdmQrCodeLayoutItem(QgsLayoutItemPicture):
-
-    def __init__(self, layout):
-        super().__init__(layout)
-
-    def type(self):
-        return STDM_QR_ITEM_TYPE
-
-    def icon(self):
-        return GuiUtils.get_icon('qrcode.png')
-
-
-class StdmQrCodeLayoutItemMetadata(QgsLayoutItemAbstractMetadata):
-
-    def __init__(self):
-        super().__init__(STDM_QR_ITEM_TYPE, QCoreApplication.translate('StdmItems', 'QR Code'))
-
-    def createItem(self, layout):
-        return StdmQrCodeLayoutItem(layout)
+from stdm.composer.custom_items.chart import StdmChartLayoutItemMetadata
+from stdm.composer.custom_items.label import StdmDataLabelLayoutItemMetadata
+from stdm.composer.custom_items.line import StdmLineLayoutItemMetadata
+from stdm.composer.custom_items.map import StdmMapLayoutItemMetadata
+from stdm.composer.custom_items.photo import StdmPhotoLayoutItemMetadata
+from stdm.composer.custom_items.qrcode import StdmQrCodeLayoutItemMetadata
+from stdm.composer.custom_items.table import StdmTableLayoutItemMetadata
 
 
 class StdmCustomLayoutItems:
